+++ src/core/utils.c
@@ -53,25 +53,25 @@ int is_greater_equal(double lhs, double rhs)
     return is_close(lhs, rhs) || lhs > rhs;
 }
 
-void *copy(void *dst, const void *src, int num, int size)
+void *copy(void *dst, const void *src, int num, size_t size)
 {
     assert(((dst && src) || num == 0) && num >= 0 && size > 0);
     if (num == 0) {
         return 0;
     }
-    return memcpy(dst, src, (size_t)num * size);
+    return memcpy(dst, src, num * size);
 }
 
-void *move(void *dst, const void *src, int num, int size)
+void *move(void *dst, const void *src, int num, size_t size)
 {
     assert(((dst && src) || num == 0) && num >= 0 && size > 0);
     if (num == 0) {
         return 0;
     }
-    return memmove(dst, src, (size_t)num * size);
+    return memmove(dst, src, num * size);
 }
 
-void sort(void *base, int num, int size, Compare *compare)
+void sort(void *base, int num, size_t size, Compare *compare)
 {
     assert((base || num == 0) && num >= 0 && size > 0 && compare);
     if (num <= 1) {
@@ -80,7 +80,7 @@ void sort(void *base, int num, int size, Compare *compare)
     qsort(base, num, size, compare);
 }
 
-int unique(void *base_, int num_, int size, Compare *compare)
+int unique(void *base_, int num_, size_t size, Compare *compare)
 {
     assert((base_ || num_ == 0) && num_ >= 0 && size > 0 && compare);
 
@@ -105,7 +105,7 @@ int unique(void *base_, int num_, int size, Compare *compare)
     return num;
 }
 
-void *search(const void *key, const void *base, int num, int size, Compare *compare)
+void *search(const void *key, const void *base, int num, size_t size, Compare *compare)
 {
     assert(key && (base || num == 0) && num >= 0 && size > 0 && compare);
     if (num == 0) {
@@ -114,7 +114,7 @@ void *search(const void *key, const void *base, int num, int size, Compare *comp
     return bsearch(key, base, num, size, compare);
 }
 
-int digitize(const void *key, const void *base_, int num, int size, Compare *compare)
+int digitize(const void *key, const void *base_, int num, size_t size, Compare *compare)
 {
     assert(key && (base_ || num == 0) && num >= 0 && size > 0 && compare);
