+++ src/teal.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <stddef.h>
+
 #define M_PI 3.14159265358979323846 /* pi */
 
 extern struct teal {
@@ -27,10 +29,10 @@ void teal_verbose(const char *fmt, ...) __attribute__((format(printf, 1, 2)));
 void teal_error(const char *fmt, ...) __attribute__((format(printf, 1, 2), noreturn));
 
 // Allocate `num * size` bytes and zero-initialize.
-void *teal_calloc(int num, int size) __attribute__((malloc));
+void *teal_calloc(int num, size_t size) __attribute__((malloc));
 
 // Reallocate `num * size` bytes, preserving existing contents.
-void *teal_realloc(void *ptr, int num, int size);
+void *teal_realloc(void *ptr, int num, size_t size);
 
 // Free memory allocated with `teal_calloc` or `teal_realloc`.
 void teal_free(void *ptr);
